test: move shared fixtures into conftest
