+++ tests/test_bitwarden_keyring.py
@@ -6,37 +6,6 @@ import pytest
 from bitwarden_keyring import backend
 
 
-# Ensure any call to a subprocess would be caught
-@pytest.fixture(autouse=True)
-def ensure_no_process(fake_process):
-    pass
-
-
-@pytest.fixture(autouse=True)
-def del_env(monkeypatch):
-    monkeypatch.delenv("BW_SESSION", raising=False)
-    yield
-
-
-@pytest.fixture
-def cred():
-    def f(
-        id: str = "id",  # noqa: A002
-        name: str | None = None,
-        username: str = "user",
-        password: str = "password",
-    ) -> backend.Credentials:
-        credentials: backend.Credentials = {
-            "id": id,
-            "login": {"username": username, "password": password},
-        }
-        if name:
-            credentials["name"] = name
-        return credentials
-
-    return f
-
-
 @pytest.mark.parametrize("path, expected", [(None, False), ("yay", True)])
 def test_bitwarden_cli_installed(path, expected):
